20250730
Basic AI and Action system in place. Still need to implement the GOAP system to determine action priorities, and create some actions to actually test if the system is working. Modified vision a bit, can now differentiate between seen and obstructed. Decreased look offset because player is short. Need to decrease eye_offset to be more accurate to the tower. Removed AI manager, we are going with an independent AI system, which will handle it's own decision-making, but will still have a slow tick rate. Actions handle actually doing things, so the tick rate doesn't have to be crazy high. For now it's 1 second, could be too long. Will need to refactor InputComponent to send signals that are no longer player specific, and any components listening to it to match.